requirements.txt
pdfplumber>=0.10.0
unidecode>=1.3.8
pyinstaller>=6.0.0
Pillow>=9.1
customtkinter>=5.2.0
packaging>=25.0